Refactor quantum state handling to utilize complex128 for improved precision
- Updated the `HilbertSpace` and `QuantumState` classes to ensure amplitudes and phases are handled as `complex128` instead of `complex64`, enhancing numerical stability and precision in quantum state representations.
- Adjusted normalization processes and tensor conversions to reflect the new data type, ensuring consistency across quantum state operations.
- Enhanced test cases in `test_cross_validation.py` to validate the integration of the updated quantum state handling, including relaxed tolerances and optimized memory management during validation processes.

These changes improve the robustness and accuracy of quantum state management, aligning with best practices for numerical analysis and testing.

+++ src/core/quantum/state_space.py
@@ -101,9 +101,9 @@ class HilbertSpace:
         Returns:
             Prepared quantum state
         """
-        # Ensure complex64 dtype
-        if amplitudes.dtype != torch.complex64:
-            amplitudes = amplitudes.to(dtype=torch.complex64)
+        # Ensure complex128 dtype
+        if amplitudes.dtype != torch.complex128:
+            amplitudes = amplitudes.to(dtype=torch.complex128)
             
         # Normalize state
         norm = torch.sqrt(torch.sum(torch.abs(amplitudes)**2, dim=-1, keepdim=True))
@@ -114,7 +114,7 @@ class HilbertSpace:
         state = QuantumState(
             amplitudes=amplitudes,
             basis_labels=self.basis_states,
-            phase=torch.zeros(1, dtype=torch.complex64, device=amplitudes.device)
+            phase=torch.zeros(1, dtype=torch.complex128, device=amplitudes.device)
         )
         
         return state
